+++ packages/types/src/index.ts
@@ -43,3 +43,5 @@ export * from './git'
 export * from './video'
 
 export * from './protocol'
+
+export * from './proxy'
