+++ packages/types/src/proxy.ts
@@ -1,0 +1,7 @@
+export type ProxyTimings = {
+  cdpRequestWillBeSentTimestamp: number
+  cdpRequestWillBeSentReceivedTimestamp: number
+  proxyRequestReceivedTimestamp: number
+  cdpLagDuration: number
+  proxyRequestCorrelationDuration: number
+}
